Completely finished education controller

+++ BackEnd/app/Http/Controllers/ProfileControllers/EducationController.php
@@ -8,8 +8,8 @@ use App\Models\Certificate;
 use Illuminate\Support\Arr;
 use Illuminate\Http\Request;
 use App\Http\Requests\EducationRequest;
-use App\Http\Requests\AddCertificateRequest;
-use App\Http\Requests\EditCertificateRequest;
+use App\Http\Requests\CertificateRequest;
+use App\Http\Resources\EducationResource;
 use App\Http\Resources\CertificateResource;
 use App\Http\Resources\CertificateCollection;
 
@@ -30,15 +30,25 @@ class EducationController extends Controller
             ], 401);
         }
 
+        // Get education
+        $education = $user->education;
+
         // Handle certificate file
         if ($request->hasFile('certificate_file')) {
+            // Check education
+            if($education != null) {
+                // Delete old certificate_file (if found)
+                $oldPath = $education->certificate_file;
+                if ($oldPath != null) {
+                    unlink(storage_path('app/'.$oldPath));
+                }
+            }
             $file = $request->file('certificate_file');
-            $path = $file->storeAs('education_files', $file->getClientOriginalName());
+            $path = $file->storeAs($user->id.'/education', $file->getClientOriginalName());
             $validated['certificate_file'] = $path;
         }
 
         // Check education
-        $education = $user->education;
         if($education == null) {
             // Create user's education
             $validated['user_id'] = $user->id;
@@ -51,19 +61,13 @@ class EducationController extends Controller
             ], 201);
         }
         else {
-            // Delete old certificate_file (if found)
-            $oldPath = $education->certificate_file;
-            if ($oldPath != null) {
-                unlink(storage_path('app/'.$oldPath));
-            }
-
             // Update user's education
             $education->update($validated);
 
             // Response
             return response()->json([
                 "message" => "Education updated",
-                "data" => $education,
+                "data" => new EducationResource($education),
             ], 200);
         }
     }
@@ -89,15 +93,7 @@ class EducationController extends Controller
         ], 200);
     }
 
-    public function ShowCertificate(Request $request, Certificate $certificate)
-    {
-        //response
-        return response()->json([
-            "data" => new CertificateResource($certificate)
-        ], 200);
-    }
-
-    public function AddCertificate(AddCertificateRequest $request)
+    public function AddCertificate(CertificateRequest $request)
     {
         // Validate request
         $validated = $request->validated();
@@ -114,10 +110,12 @@ class EducationController extends Controller
 
         // Handle certificate file
         if ($request->hasFile('file')) {
-            $avatarPath = $request->file('file')->store('files', 'public');
-            $validated['file'] = $avatarPath;
+            $file = $request->file('file');
+            $path = $file->storeAs($user->id.'/certificates', $file->getClientOriginalName());
+            $validated['file'] = $path;
         }
 
+        // Create certificate
         $validated['user_id'] = $user->id;
         $certificate = Certificate::create($validated);
 
@@ -128,7 +126,7 @@ class EducationController extends Controller
         ], 201);
     }
 
-    public function EditCertificate(EditCertificateRequest $request, Certificate $certificate)
+    public function EditCertificate(CertificateRequest $request, Certificate $certificate)
     {
         // Validate request
         $validated = $request->validated();
@@ -143,11 +141,28 @@ class EducationController extends Controller
             ], 401);
         }
 
+        // Check if certificate belongs to user
+        if($user->id != $certificate->user_id) {
+            return response()->json([
+                'user' => 'Invalid user'
+            ], 401);
+        }
+
         // Handle certificate file
         if ($request->hasFile('file')) {
-            $avatarPath = $request->file('file')->store('files', 'public');
-            $validated['file'] = $avatarPath;
+            // Delete old certificate_file (if found)
+            $oldPath = $certificate->file;
+            if ($oldPath != null) {
+                unlink(storage_path('app/'.$oldPath));
+            }
+
+            // Add new certificate_file
+            $file = $request->file('file');
+            $path = $file->storeAs($user->id.'/certificates', $file->getClientOriginalName());
+            $validated['file'] = $path;
         }
+
+        // Update user's certificate
         $certificate->update($validated);
 
         // Response
@@ -159,6 +174,30 @@ class EducationController extends Controller
 
     public function DeleteCertificate(Request $request, Certificate $certificate)
     {
+        // Get User
+        $user = auth()->user();
+
+        // Check user
+        if ($user == null) {
+            return response()->json([
+                'user' => 'Invalid user'
+            ], 401);
+        }
+
+        // Check if certificate belongs to user
+        if($user->id != $certificate->user_id) {
+            return response()->json([
+                'user' => 'Invalid user'
+            ], 401);
+        }
+
+        // Delete old certificate_file (if found)
+        $oldPath = $certificate->file;
+        if ($oldPath != null) {
+            unlink(storage_path('app/'.$oldPath));
+        }
+
+        // Delete certificate
         $certificate->delete();
 
         return response()->json([
